refactoring 'backend' imports
