+++ frontend/src/components/Games/GameContainer.tsx
@@ -1,14 +1,12 @@
 import React from 'react'
-import { Modal, ModalCloseButton, ModalContent, ModalOverlay, useDisclosure} from "@chakra-ui/react";
-import MenuItem from "@material-ui/core/MenuItem";
-import Typography from "@material-ui/core/Typography";
+import { Button, Modal, ModalCloseButton, ModalContent, ModalOverlay, useDisclosure} from "@chakra-ui/react";
 import TicTacToeDisplay from "./GameDisplays/TicTacToe/TicTacToeDisplay";
 import TTLDisplay from "./GameDisplays/TTLDisplay";
 import HangmanDisplay from "./GameDisplays/Hangman/HangmanDisplay";
-import GameController from "../../../../services/roomService/src/games/GameController";
-import TTLGame from "../../../../services/roomService/src/games/TTLGame";
-import TicTacToeGame from "../../../../services/roomService/src/games/TicTacToeGame";
-import HangmanGame from "../../../../services/roomService/src/games/HangmanGame";
+import GameController from "./gamesService/GameController";
+import TTLGame from "./gamesService/TTLGame";
+import TicTacToeGame from "./gamesService/TicTacToeGame";
+import HangmanGame from "./gamesService/HangmanGame";
 
 interface GameContainerProps {
   gameType: string;
